requirements.txt
# Password hashing library for secure user authentication
bcrypt

# Data manipulation and analysis library
pandas

# Web application framework for building interactive dashboards
streamlit

# SQLite database driver (binary distribution for compatibility)
sqlite3-binary

# Date and time parsing utilities
python-dateutil

# Google Generative AI SDK for Gemini API integration
google-generativeai

# Interactive data visualization library for charts and graphs
plotly

# Image processing library (PIL/Pillow) for profile picture handling
Pillow

# Numerical computing library for array operations and mathematical functions
numpy
